refactor: hyprtester/plugin
Refactored to make it easier to add new things in the future.
